% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_functions.R
\name{commission}
\alias{commission}
\title{Parameter commission}
\usage{
commission(premi, presentase_komisi)
}
\arguments{
\item{premi}{nilai premi pertanggungan asuransi pada periode tertentu}

\item{presentase_komisi}{komisi yang akan diberikan kepada tertanggung}
}
\value{
nilai komisi untuk tertanggung
}
\description{
Parameter commission
}
\examples{
commission1 <- commission(200000*0.25)
commission2 <- commission(5000*0.35)
}
