% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_functions.R
\name{loss_ratio}
\alias{loss_ratio}
\title{parameter loss_ratio}
\usage{
loss_ratio(premi, klaim)
}
\arguments{
\item{premi}{nilai premi}

\item{klaim}{nilai klaim}
}
\value{
loss ratio dari pertanggungan asuransi
}
\description{
parameter loss_ratio
}
\examples{
loss_ratio1 <-(50000,10000)
loss_ratio2 <-(2000000,2500000)
}
