% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_functions.R
\name{premium}
\alias{premium}
\title{Parameter Premium}
\usage{
premium(coverage, rate, periode = 1)
}
\arguments{
\item{coverage}{nilai pertanggungan}

\item{rate}{nilai rate atas pertanggungan}

\item{periode}{berapa lama pertanggungan tersebut di asuransikan (tahun)}
}
\value{
nilai premi
}
\description{
Parameter Premium
}
\examples{
premium1<- premium(1000000,0.019,0.8)
premium2 <- premium(5000000,0.0433)
}
